Refactor zwe components install for dry run

+++ bin/commands/components/install/component-script/index.ts
@@ -18,7 +18,7 @@ import * as component from '../../../../libs/component';
 import * as varlib from '../../../../libs/var';
 import { PathAPI as pathoid } from '../../../../libs/pathoid';
 
-export function execute(componentName: string, zisPluginDatasets: string[]) {
+export function execute(componentName: string, zisPluginDatasets: string[], dryRun?: boolean) {
   common.requireZoweYaml();
   const ZOWE_CONFIG=config.getZoweConfig();
   // read extensionDirectory
@@ -35,28 +35,31 @@ export function execute(componentName: string, zisPluginDatasets: string[]) {
     common.printMessage(`Process ${installScript} defined in manifest commands.install:`);
     const scriptPath = pathoid.join(targetDir, componentName, installScript);
     const componentRoot = pathoid.join(targetDir, componentName);
-    // run commands
-    const result = shell.execOutSync('sh', '-c', `. ${ZOWE_CONFIG.zowe.runtimeDirectory}/bin/libs/configmgr-index.sh && cd ${componentRoot} && . ${scriptPath} ; export rc=$? ; export -p`);
-    if (result.rc==0) {
-      varlib.getEnvironmentExports(result.out, true);
-      const outLines = result.out.split('\n');
-      common.printFormattedInfo("ZWELS", "zwe-components-install-process-hook", `- commands.install output from ${componentName} is:`);
-      common.printMessage(outLines.filter(line => !line.startsWith('export ')).join('\n'));
-      common.printFormattedDebug("ZWELS", "zwe-components-install-process-hook", outLines.filter(line => line.startsWith('export ')).join('\n'));
-    } else {
-      common.printError(`install script ended with error, rc=${result.rc}`);
-      if (result.out) {
+
+    const command = `. ${ZOWE_CONFIG.zowe.runtimeDirectory}/bin/libs/configmgr-index.sh && cd ${componentRoot} && . ${scriptPath} ; export rc=$? ; export -p`;
+    common.printFormattedInfo("ZWELS", "zwe-components-install-process-hook", `Running component install script "${scriptPath}"`);
+    common.printFormattedInfo("ZWELS", "zwe-components-install-process-hook", `Running command ${command}`);
+    if (!dryRun) {
+      // run commands
+      const result = shell.execOutSync('sh', '-c', command);
+      if (result.rc==0) {
+        varlib.getEnvironmentExports(result.out, true);
         const outLines = result.out.split('\n');
         common.printFormattedInfo("ZWELS", "zwe-components-install-process-hook", `- commands.install output from ${componentName} is:`);
         common.printMessage(outLines.filter(line => !line.startsWith('export ')).join('\n'));
+        common.printFormattedDebug("ZWELS", "zwe-components-install-process-hook", outLines.filter(line => line.startsWith('export ')).join('\n'));
+      } else {
+        common.printError(`install script ended with error, rc=${result.rc}`);
+        if (result.out) {
+          const outLines = result.out.split('\n');
+          common.printFormattedInfo("ZWELS", "zwe-components-install-process-hook", `- commands.install output from ${componentName} is:`);
+          common.printMessage(outLines.filter(line => !line.startsWith('export ')).join('\n'));
+        }
+        std.exit(result.rc);
       }
-      std.exit(result.rc);
     }
 
   } else {
     common.printDebug(`Module ${componentName} does not have commands.install defined.`);
   }
-
-  component.processZssPluginInstall(componentDir);
-  component.processZisPluginInstall(componentDir, zisPluginDatasets);
 }
